requirements.txt
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.15
django-allauth==0.42.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-import-export==2.3.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pep8==1.7.1
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
tablib==2.0.0
toml==0.10.1
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
